tests/wpt/web-platform-tests/svg/geometry/parsing/ry-computed.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="8000px">
  <title>SVG Geometry Properties: getComputedStyle().ry</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/geometry.html#RY"/>
  </metadata>
  <g id="target"></g>
  <style>
    #target {
      font-size: 40px;
    }
  </style>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/computed-testcommon.js"/>
  <script><![CDATA[

test_computed_value("ry", "auto");

test_computed_value("ry", "10px");
test_computed_value("ry", "0.5em", "20px");
test_computed_value("ry", "calc(10px + 0.5em)", "30px");
test_computed_value("ry", "calc(10px - 0.5em)", "0px");
test_computed_value("ry", "40%");
test_computed_value("ry", "calc(50% + 60px)");

  ]]></script>
</svg>
